Move client auth cert tests to client_cert_verifier.rs
